+++ tests/max_encoded_len_ui/unsupported_field.rs
@@ -1,0 +1,12 @@
+use jam_codec::{Encode, MaxEncodedLen};
+
+#[derive(Encode)]
+struct NotMel;
+
+#[derive(Encode, MaxEncodedLen)]
+struct UnsupportedField {
+	mel: u32,
+	not_mel: NotMel,
+}
+
+fn main() {}